change .dot() to @

+++ HW/HW03/HW03_py/HW03_1.py
@@ -6,8 +6,8 @@ from data_generator import UnivariateGaussian, PolynomialBasis
 choose_type = int(input("0 for Univariate Gaussian, 1 for Polynomail Basis Linear Model： "))
 
 if choose_type == 0:
-    mean = int(input("Mean： "))
-    var = int(input("Variance： "))
+    mean = float(input("Mean： "))
+    var = float(input("Variance： "))
 
     points = []
     for i in range(10000):
@@ -18,7 +18,7 @@ elif choose_type == 1:
     w = []
     for i in range(n):
         w.append(float(input("Coefficient of x^" + str(i) + "： ")))
-    a = int(input("Variance： "))
+    a = float(input("Variance： "))
 
     points = []
     for i in range(10000):
